one_thing

+++ 17-lecture_switch/different_body_q7.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-     int BMI;
+     float BMI;
      int average;
      int weight;
      float height;
